src/test/java/MasterControlTest.java
import org.junit.jupiter.api.Test;

import java.util.List;

import static org.junit.Assert.assertEquals;

public class MasterControlTest {
    List<String> input;
    Bank bank = new Bank();
    CommandValidator commandValidator = new CommandValidator(bank);
    CommandProcessor commandProcessor = new CommandProcessor(bank);
    CommandStorage commandStorage = new CommandStorage();
    MasterControl masterControl = new MasterControl(commandValidator, commandProcessor, commandStorage);

    @Test
    void sample(){
        input.add("Create savings 12345678 0.6");
        input.add("Deposit 12345678 700");
        input.add("Deposit 12345678 5000");
        input.add("creAte cHecKing 98765432 0.01");
        input.add("Deposit 98765432 300");
        input.add("Transfer 98765432 12345678");
        input.add("Pass 1");
        input.add("Create cd 23456789 1.2 2000");
        List<String> actual = masterControl.start(input);

        assertEquals(5, actual.size());
        assertEquals("Savings 12345678 1000.50 0.60", actual.get(0));
        assertEquals("Deposit 12345678 700", actual.get(1));
        assertEquals("Transfer 98765432 12345678 300", actual.get(2));
        assertEquals("cd 23456789 2000.00 1.20", actual.get(3));
        assertEquals("Deposit 12345678 5000", actual.get(4));
    }
}
